// src/app/shared/models/header.model.ts
export enum Currencies {
  RUB = 'RUB',
  EUR = 'EUR',
  USD = 'USD',
}
export enum DateFormat {
  DMY = 'DD/MM/YYYY',
  MDY = 'MM/DD/YYYY',
  YDM = 'YYYY/DD/MM',
  YMD = 'YYYY/MM/DD',
}

export enum HeaderActions {
  UPDATE_CURRENCY = 'UPDATE_CURRENCY',
  UPDATE_DATE_FORMAT = 'UPDATE_DATE_FORMAT',
  SHOW_PROGRESS = 'SHOW_PROGRESS',
}